pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lyricforge"
version = "0.1.0"
description = "Analyze English song lyrics and generate flashcards for language learning"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "lyricsgenius>=3.0.0",
    "ollama>=0.4.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
lyricforge = "lyricforge.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["lyricforge"]
